Capture Label prenormalization ph02 and ph 03
Capture Label reverted to state before normalization of percentage of phase 2 and phase 3. (Reason: Poor CNN performance after normalization)

+++ Project/CNN/CNN/CNN/Utilities.py
@@ -224,7 +224,7 @@ def readMinMaxFromCSV(filepath):
 			
 			cnt_cat = 0
 			for item in cat:
-				cat[cnt_cat] = int(item)
+				cat[cnt_cat] = float(item)
 				cnt_cat = cnt_cat + 1
 			cat = np.asarray(cat)
